Add an option to turn task mode on
Added a switch toggle turning task mode off and on to listview.html. Updated the JS die rolling function accordingly. The RollHistory list in the database is emptied when the task mode is turned off, and when the user rolls all the items from the list and keeps rolling - the list is "reset", and the list item cards are not greyed out anymore.

That's quite a big commit. Sorry about that; I forgot to commit in the process.